+++ docs/src/main/paradox/release-notes.md
@@ -15,6 +15,7 @@
 * In `rf_ipython`, improved rendering of dataframe binary contents with PNG preamble.
 * Throw an `IllegalArgumentException` when attempting to apply a mask to a `Tile` whose `CellType` has no NoData defined. ([#409](https://github.com/locationtech/rasterframes/issues/384))
 * Add `rf_agg_approx-quantiles` function to compute cell quantiles across an entire column.
+* Add `rf_local_min`, `rf_local_max`, and `rf_local_clip` functions.
 
 ### 0.8.4